Change type of dot_product to int64_t

+++ tasks/word2vec/word2vec.cpp
@@ -2,8 +2,8 @@
 
 #include <vector>
 
-int DotProduct(std::vector<int> vector_1, std::vector<int> vector_2) {
-    int dot_product = 0;
+int64_t DotProduct(std::vector<int> vector_1, std::vector<int> vector_2) {
+    int64_t dot_product = 0;
     for (size_t i = 0; i < vector_1.size(); ++i) {
         dot_product += vector_1[i] * vector_2[i];
     }
@@ -16,7 +16,7 @@ std::vector<std::string> FindClosestWords(const std::vector<std::string>& words,
         return {};
     }
     std::vector<std::string> closest_words = {words[1]};
-    int current_max_dot_product = DotProduct(vectors[0], vectors[1]);
+    int64_t current_max_dot_product = DotProduct(vectors[0], vectors[1]);
     for (size_t i = 2; i < vectors.size(); ++i) {
         int current_dot_product = DotProduct(vectors[0], vectors[i]);
         if (current_dot_product > current_max_dot_product) {
